re-index tests
