+++ src/test/packages/03-deprecated-component-scripts/zarf.yaml
